requirements.txt
accelerate==1.12.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
APScheduler==3.11.1
asyncio==4.0.0
attrs==25.4.0
audioread==3.1.0
av==16.0.1
bitsandbytes==0.48.2
black==25.11.0
brotli==1.2.0
budoux==0.7.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cmake==4.2.0
cn2an==0.5.23
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.10.1
contourpy==1.3.3
coverage==7.12.0
cycler==0.12.1
decorator==5.2.1
diskcache==5.6.3
Distance==0.1.3
einops==0.8.1
einx==0.3.0
fast-langdetect==1.0.0
fastapi==0.122.0
fasttext-predict==0.9.2.4
ffmpeg-python==0.2.0
ffmpy==1.0.0
filelock==3.20.0
flake8==7.3.0
flatbuffers==25.9.23
fonttools==4.61.0
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2025.10.0
future==1.0.0
g2p-en==2.1.0
gradio==3.46.0
gradio_client==0.5.3
groovy==0.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
importlib_resources==6.5.2
inflect==7.5.0
iniconfig==2.3.0
jieba==0.42.1
jieba_fast==0.53
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lazy_loader==0.4
librosa==0.11.0
librt==0.6.3
lightning-utilities==0.15.2
llama_cpp_python==0.3.16
llvmlite==0.45.1
loguru==0.7.3
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.7
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
mypy==1.19.0
mypy_extensions==1.1.0
narwhals==2.12.0
networkx==3.6
ninja==1.13.0
nltk==3.9.2
numba==0.62.1
numpy==1.26.4
nvidia-ml-py==13.580.82
onnxruntime==1.23.2
opencc-python-reimplemented==0.1.7
orjson==3.11.4
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
peft==0.12.0
pillow==10.4.0
platformdirs==4.5.0
pluggy==1.6.0
pooch==1.8.2
praat-parselmouth==0.4.7
proces==0.1.7
propcache==0.4.1
protobuf==6.33.1
psutil==7.1.3
pyAudioAnalysis==0.3.14
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydub==0.25.1
pyflakes==3.4.0
Pygments==2.19.2
pynvml==13.0.1
pyparsing==3.2.5
pypinyin==0.55.0
pyreadline3==3.5.4
pytest==9.0.1
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytokens==0.3.0
pytorch-lightning==2.6.0
pytz==2025.2
pywin32==311
PyYAML==6.0.3
qwen-vl-utils==0.0.14
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
resampy==0.4.3
rich==14.2.0
robust-downloader==0.0.2
rpds-py==0.29.0
ruff==0.14.7
safehttpx==0.1.7
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
semantic-version==2.10.0
sentencepiece==0.2.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sounddevice==0.5.3
soundfile==0.13.1
soxr==1.0.0
SpeechRecognition==3.14.4
split-lang==2.1.1
starlette==0.50.0
structlog==25.5.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.4
tomlkit==0.12.0
torch==2.9.1+cu128
torchaudio==2.9.1
torchcodec==0.8.1
torchmetrics==1.8.2
tqdm==4.67.1
transformers==4.48.0
typeguard==4.4.4
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.38.0
watchdog==6.0.0
websockets==11.0.3
win32_setctime==1.2.0
wordsegment==1.3.1
x-transformers==2.11.23
yarl==1.22.0
